fix: marker, 계절 icon ul 처리, service 폴더 생성
