+++ src/components/Marker.tsx
@@ -8,14 +8,16 @@ type Props = {
 };
 export default function Marker({ bookmarks }: Props) {
   return (
-    <div>
+    <ul>
       {bookmarks.map(bookmark => (
-        <MarkerF
-          key={bookmark.lat}
-          position={{ lat: bookmark.lat, lng: bookmark.lng }}
-          onLoad={() => console.log("Marker Loaded")}
-        />
+        <li>
+          <MarkerF
+            key={bookmark.lat}
+            position={{ lat: bookmark.lat, lng: bookmark.lng }}
+            onLoad={() => console.log("Marker Loaded")}
+          />
+        </li>
       ))}
-    </div>
+    </ul>
   );
 }
