+++ src/components/ui/icons/WinterIcon.tsx
@@ -1,5 +1,5 @@
 import { BsSnow2 } from "react-icons/bs";
 
 export default function WinterIcon() {
-  return <BsSnow2 className="w-full h-full  text-cyan-500 p-2" />;
+  return <BsSnow2 className="w-full h-full text-cyan-500 p-2" />;
 }
